refactor: folders and files
