+++ components.json
@@ -11,7 +11,7 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/views/components",
+    "components": "@/components",
     "utils": "@/lib/utils"
   }
 }
