+++ src/app.tsx
@@ -1,6 +1,6 @@
+import { Toaster } from './components/ui/sonner'
 import { ContactsProvider } from './contexts/contacts-context'
 import { Router } from './router'
-import { Toaster } from './views/components/ui/sonner'
 
 import './globals.css'
