+++ src/components/ui/alert-dialog.tsx
@@ -1,8 +1,8 @@
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog'
 import * as React from 'react'
 
+import { buttonVariants } from '@/components/ui/button'
 import { cn } from '@/lib/utils'
-import { buttonVariants } from '@/views/components/ui/button'
 
 const AlertDialog = AlertDialogPrimitive.Root
