+++ src/pages/edit-contact.tsx
@@ -3,11 +3,10 @@ import { useCallback, useEffect, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import { toast } from 'sonner'
 
+import { ContactForm, TContactFormSchema } from '@/components/contact-form'
+import { Button } from '@/components/ui/button'
+import { Spinner } from '@/components/ui/spinner'
 import { IContact, useContacts } from '@/contexts/contacts-context'
-
-import { ContactForm, TContactFormSchema } from '../components/contact-form'
-import { Button } from '../components/ui/button'
-import { Spinner } from '../components/ui/spinner'
 
 export function EditContact() {
   const [currentContact, setCurrentContact] = useState<
