+++ src/pages/home/index.tsx
@@ -2,11 +2,10 @@ import { Contact } from 'lucide-react'
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 
+import { Button } from '@/components/ui/button'
+import { Spinner } from '@/components/ui/spinner'
 import { useContacts } from '@/contexts/contacts-context'
 import { isMobile } from '@/lib/utils'
-import { Spinner } from '@/views/components/ui/spinner'
-
-import { Button } from '../../components/ui/button'
 
 import { ContactCard } from './components/contact-card'
 import { Filters } from './components/filters'
