+++ src/pages/new-contact.tsx
@@ -2,11 +2,10 @@ import { Check, CircleChevronLeft } from 'lucide-react'
 import { useNavigate } from 'react-router-dom'
 import { toast } from 'sonner'
 
+import { ContactForm, TContactFormSchema } from '@/components/contact-form'
+import { Button } from '@/components/ui/button'
+import { Spinner } from '@/components/ui/spinner'
 import { useContacts } from '@/contexts/contacts-context'
-
-import { ContactForm, TContactFormSchema } from '../components/contact-form'
-import { Button } from '../components/ui/button'
-import { Spinner } from '../components/ui/spinner'
 
 export function NewContact() {
   const navigate = useNavigate()
