+++ src/pages/not-found.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom'
 
-import { Button } from '../components/ui/button'
+import { Button } from '@/components/ui/button'
 
 export function NotFound() {
   const navigate = useNavigate()
