+++ src/router.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
-import { DefaultLayout } from './views/layouts/default-layout'
-import { EditContact } from './views/pages/edit-contact'
-import { Home } from './views/pages/home'
-import { NewContact } from './views/pages/new-contact'
-import { NotFound } from './views/pages/not-found'
+import { DefaultLayout } from './layouts/default-layout'
+import { EditContact } from './pages/edit-contact'
+import { Home } from './pages/home'
+import { NewContact } from './pages/new-contact'
+import { NotFound } from './pages/not-found'
 
 export function Router() {
   return (
